improve and evaluate refactor
