+++ streamlit/pages/6_🏋️‍♀️_Improve & Evaluate.py
@@ -35,206 +35,6 @@ if st.sidebar.button('Clear Cache'):
     clear_all_caches()
     st.sidebar.success('Cache cleared!')
 
-def fetch_result_data(lesson_plan_id, prompt_id, result):
-    try:
-        conn = get_db_connection()
-        query = """
-        SELECT 
-            r.prompt_id, r.lesson_plan_id, r.result, r.justification
-        FROM public.m_results r
-        WHERE r.lesson_plan_id = %s 
-        AND r.prompt_id = %s 
-        AND CAST(r.result AS numeric) = %s
-        AND r.status = 'SUCCESS'
-        ORDER BY r.result ASC
-        LIMIT 1
-        """
-        df = pd.read_sql_query(query, conn, params=(lesson_plan_id, prompt_id, result))
-        conn.close()
-        return df
-    except Exception as e:
-        print(f"An error occurred while fetching result data: {e}")
-        return None
-
-
-def fetch_final_data(lesson_plan_id, prompt_id, experiment_id):
-    try:
-        conn = get_db_connection()
-        query = """
-        SELECT 
-            r.result, r.justification
-        FROM public.m_results r
-        WHERE r.lesson_plan_id = %s 
-        AND r.prompt_id = %s 
-        AND r.experiment_id = %s
-        """
-        df = pd.read_sql_query(query, conn, params=(lesson_plan_id, prompt_id, experiment_id))
-        conn.close()
-        return df
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return None
-    
-def delete_created_sample(sample_id):
-    try:
-        conn = get_db_connection()
-        query = """
-        DELETE FROM public.m_samples
-        WHERE id = %s
-        """
-        cur = conn.cursor()
-        cur.execute(query, (sample_id,))
-        conn.commit()
-        conn.close()
-        return True
-    except Exception as e:
-        print(f"An error occurred while deleting the sample: {e}")
-        return False
-def delete_lesson_plans_from_sample_lesson_plans(sample_id):
-    try:
-        conn = get_db_connection()
-        query = """
-        DELETE FROM public.m_sample_lesson_plans
-        WHERE sample_id = %s
-        """
-        cur = conn.cursor()
-        cur.execute(query, (sample_id,))
-        conn.commit()
-        conn.close()
-        return True
-    except Exception as e:
-        print(f"An error occurred while deleting the sample lesson plans: {e}")
-        return False
-
-def run_agent_openai_inference(prompt, llm_model, llm_model_temp, timeout=150):
-        client = OpenAI( api_key= os.environ.get("OPENAI_API_KEY"), timeout=timeout)
-
-        
-        try:
-            start_time = time.time()
-
-            response = client.chat.completions.create(
-                model=llm_model,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=llm_model_temp,
-                top_p=1,
-                frequency_penalty=0,
-                presence_penalty=0,
-            )
-            message = response.choices[0].message.content
-            # print(message)
-            end_time = time.time()
-
-            # Calculate the duration
-            duration = end_time - start_time
-            cleaned_content, status = clean_response(message)
-            return {
-                "response": cleaned_content,
-                "response_time": duration,
-            }
-
-        except Exception as e:
-            log_message("error", f"Unexpected error during inference: {e}")
-            return {
-                "response": {
-                    "result": None,
-                    "justification": f"An error occurred: {e}",
-                },
-                "status": "FAILURE",
-                "response_time": duration,
-            }
-
-def fetch_lesson_plan_json(lesson_plan_id):
-    try:
-        conn = get_db_connection()
-        query = """
-        SELECT json
-        FROM public.lesson_plans 
-        WHERE id = %s
-        """
-        df = pd.read_sql_query(query, conn, params=(lesson_plan_id,))
-        conn.close()
-        return df
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return None
-
-
-def calculate_success_failure_rate(df):
-    
-    df = df.groupby(['lesson_plan_id', 'generation_details', 'prompt_id']).agg({
-        'min_result': 'min',
-        'max_result': 'max',
-        'justification_count': 'sum',
-        'score_1_count': 'sum',
-        'score_2_count': 'sum',
-        'score_3_count': 'sum',
-        'score_4_count': 'sum',
-        'score_5_count': 'sum',
-        'prompt_title': lambda x: ', '.join(x)  # Concatenate prompt titles as a comma-separated string
-    }).reset_index()
-
-
-    df['stellar_success_rate'] = (df['score_5_count'] / df['justification_count']) * 100
-    df['catastrophic_fail_rate'] = (df['score_1_count'] / df['justification_count']) * 100
-
-    df = df[['lesson_plan_id', 'generation_details','prompt_title', 'prompt_id' , 'stellar_success_rate', 'catastrophic_fail_rate']]
-
-    #rename 'prompt_title' to 'prompt_titles'
-    # display_df.rename(columns={'prompt_title': 'prompt_titles'}, inplace=True)
-    df['overall_fail_score'] = (100 - df['stellar_success_rate']) + df['catastrophic_fail_rate']
-
-    return df
-
-
-def fetch_prompt_objectives_desc():
-    try:
-        conn = get_db_connection()
-        query = """
-        SELECT id, objective_desc
-        FROM public.m_prompts
-        WHERE output_format = 'Score'
-        """
-        df = pd.read_sql_query(query, conn)
-        conn.close()
-        return df
-    except Exception as e:
-        print(f"An error occurred while fetching prompt details: {e}")
-        return None
-
-
-def fetch_bad_lesson_plans(selected_prompt_ids):
-    try:
-        conn = get_db_connection()  
-        query = """SELECT 
-                    r.prompt_id, 
-                    r.lesson_plan_id, 
-                    lp.generation_details,
-                    p.prompt_title,
-                    min(CAST(r.result AS numeric)) AS min_result, 
-                    max(CAST(r.result AS numeric)) AS max_result,
-                    count(r.justification) AS justification_count,
-                    COUNT(CASE WHEN CAST(r.result AS numeric) = 1 THEN 1 END) AS score_1_count,
-                    COUNT(CASE WHEN CAST(r.result AS numeric) = 2 THEN 1 END) AS score_2_count,
-                    COUNT(CASE WHEN CAST(r.result AS numeric) = 3 THEN 1 END) AS score_3_count,
-                    COUNT(CASE WHEN CAST(r.result AS numeric) = 4 THEN 1 END) AS score_4_count, 
-                    COUNT(CASE WHEN CAST(r.result AS numeric) = 5 THEN 1 END) AS score_5_count
-                FROM public.m_results r
-                INNER JOIN m_prompts p ON p.id = r.prompt_id
-                INNER JOIN lesson_plans lp ON lp.id = r.lesson_plan_id
-                WHERE r.status = 'SUCCESS' AND r.result ~ '^[0-9\\.]+$' AND p.output_format = 'Score' 
-                AND p.prompt_title <> 'Answers Are Minimally Different'
-                AND r.prompt_id IN %s
-                GROUP BY r.lesson_plan_id, r.prompt_id, p.prompt_title, lp.generation_details
-                ORDER BY lesson_plan_id DESC, justification_count DESC, max_result ASC;"""
-        df = pd.read_sql_query(query, conn, params=(selected_prompt_ids,))
-        conn.close()
-        return df
-    
-    except Exception as e:
-        print(f"An error occurred while fetching lesson plans: {e}")
-        return None
-
 
 #set page logo
